modules/location_endpoint/requirements.txt
kafka-python
grpcio-tools==1.46.5
Protobuf==3.19.5


